Add #to, #append, and #dropLast.
Binder#to applies transforms from a target binder after the
transforms in the current binder, returning a new Binder. A new
Binder is returned to avoid modifying either of the original
Binders.

Binder#append is the same as Binder#insert with the index of the
last argument + 1.

Binder#dropLast is the same as Binder#drop with the index of the
last argument + 1.