First attempt to work around slow asCollector
The logic here duplicates what the slow asCollector does at a
MethodHandle level, collecting all args into an Object[] and then
arraycopying the contents into a new instance of the array we
actually want. This does not appear to be faster than the slow
asCollector, probably because the intermediate Object[] is still
not being eliminated.